Functionally working, although still mathematically incorrect

+++ pypwdg/utils/sparseutils.py
@@ -15,4 +15,4 @@ def getintptr(indices, n):
 def sparseindex(rows, cols, m, n=None):
     """ Return a csr matrix with a one at all the points given in rows and cols """
     if n is None: n = m
-    return sparse.csr_matrix((numpy.ones(len(rows)), cols, getintptr(rows, m)),shape=(m,n))
+    return sparse.csr_matrix((numpy.ones(len(rows),dtype=int), cols, getintptr(rows, m)),shape=(m,n))
